+++ minecraft-plugin/src/main/java/com/pvchat/proximityvoicechat/plugin/distanceMatrix/PlayerVolumeData.java
@@ -5,9 +5,9 @@ import com.pvchat.proximityvoicechat.plugin.DiscordUserID;
 public class PlayerVolumeData {
  private DiscordUserID player1;
  private DiscordUserID player2;
- private long volumeLevel;
+ private int volumeLevel;
 
-    public PlayerVolumeData(DiscordUserID player1ID, DiscordUserID player2ID, long volumeLevel) {
+    public PlayerVolumeData(DiscordUserID player1ID, DiscordUserID player2ID, int volumeLevel) {
         this.player1 = player1ID;
         this.player2 = player2ID;
         this.volumeLevel = volumeLevel;
@@ -29,11 +29,11 @@ public class PlayerVolumeData {
         this.player2 = player2;
     }
 
-    public long getVolumeLevel() {
+    public int getVolumeLevel() {
         return volumeLevel;
     }
 
-    public void setVolumeLevel(long volumeLevel) {
+    public void setVolumeLevel(int volumeLevel) {
         this.volumeLevel = volumeLevel;
     }
 }
